module2/text101/requirements.txt
python-dotenv #bulut servislere erişmek için
openai #SDK
streamlit #web uygulamamızı ayağa kaldırmak için 
google-generativeai #SDK
anthropic #SDK
cohere #SDK


